Uplift complex type ops back into typeck so we can call them locally
